+++ match/src/main/java/com/itranswarp/crypto/match/Tick.java
@@ -1,14 +1,16 @@
 package com.itranswarp.crypto.match;
 
+import java.math.BigDecimal;
+
 public class Tick {
 
 	public final long time;
 
-	public final long price;
+	public final BigDecimal price;
 
-	public final long amount;
+	public final BigDecimal amount;
 
-	public Tick(long time, long price, long amount) {
+	public Tick(long time, BigDecimal price, BigDecimal amount) {
 		this.time = time;
 		this.price = price;
 		this.amount = amount;
@@ -16,6 +18,6 @@ public class Tick {
 
 	@Override
 	public String toString() {
-		return String.format("Tick(time=%d, price=%d, amount=%d)", time, price, amount);
+		return String.format("Tick(time=%d, price=%.2f, amount=%.4f)", time, price, amount);
 	}
 }
